More testing... vec_shl. Interestingly, you really have to assign a stringified value to the variable or else it will be treated as a reference to the data it was initialised from!
Changed default values as I've gotten to the stage where I need
to print out data structures in order to be able to debug the
program.

Big bug fixed. I assumed that vec working on bits looked at the
most significant bit first. It's the other way around. Wrote
a wrapper function to make it work the way I want.

I think that I had a bug in how I was truncating @arows. I
changed from a map {} version to a for() {} version.

Lots of debug code.