Make change relating to code. See last commit log message. Port more code.
Right now, it takes ~2.7s to simply call the constructor for
the object, without even calling pivot once. While this cost
can be amortised over the time spent sending a large file
across the network, it suggests that OpenCL will not bring
any benefit when transferring small files. The compiler on
a Raspberry Pi is likely to be many times slower, too.

Also, I note that if I increase WORKSIZE to 12, the declaration
of sym[WORKSIZE] fails as it did with code[ALPHA]. Therefore,
it seems that I should convert the code to working with a global
sym working variable, as I did with code[]. At least this has an
upside, though, since threads only work on a small section of
the string, and we can use the output symbol variable as our
working variable.